move most of event entry styling to own component
