+++ src/main/js/events/PluginCenterNotAvailableEvent.tsx
@@ -29,30 +29,11 @@ import { useTranslation } from "react-i18next";
 import styled from "styled-components";
 import classNames from "classnames";
 import { Link as ReactLink } from "react-router-dom";
+import MyEventEntry from "./MyEventEntry";
 
 type PluginCenterEventType = MyEventType & {
   date: Date;
 };
-
-const FlexFullHeight = styled.div`
-  flex-direction: column;
-  justify-content: space-around;
-  align-self: stretch;
-`;
-
-const ContentLeft = styled.div`
-  margin-bottom: 0 !important;
-  overflow: hidden;
-`;
-
-const ContentRight = styled.div`
-  margin-left: auto;
-  align-items: start;
-`;
-
-const CenteredItems = styled.div`
-  align-items: center;
-`;
 
 const Icon = styled.i`
   width: 2.5rem;
@@ -60,37 +41,15 @@ const Icon = styled.i`
   align-self: center;
 `;
 
-const StyledLink = styled(ReactLink)`
-  color: inherit;
-  :hover {
-    color: #33b2e8 !important;
-  }
-`;
-
 const PluginCenterNotAvailableEvent: MyEventComponent<PluginCenterEventType> = ({ event }) => {
   const [t] = useTranslation("plugins");
-
   const link = "/admin/plugins/available/";
-
-  return (
-    <StyledLink to={link}>
-      <div className={"media"}>
-        <Icon className="fas fa-puzzle-piece fa-2x media-left" />
-        <FlexFullHeight className={classNames("media-content", "text-box", "is-flex")}>
-          <CenteredItems className="is-flex">
-            <ContentLeft className="content">
-              <strong className="is-marginless">
-                {t("scm-landingpage-plugin.myevents.pluginCenterNotAvailable.title")}
-              </strong>
-            </ContentLeft>
-            <ContentRight>
-              <DateFromNow date={event.date} />
-            </ContentRight>
-          </CenteredItems>
-        </FlexFullHeight>
-      </div>
-    </StyledLink>
+  const icon = <Icon className="fas fa-puzzle-piece fa-2x media-left" />;
+  const contentLeft = (
+    <strong className="is-marginless">{t("scm-landingpage-plugin.myevents.pluginCenterNotAvailable.title")}</strong>
   );
+
+  return <MyEventEntry link={link} icon={icon} contentLeft={contentLeft} date={event.date} />;
 };
 
 PluginCenterNotAvailableEvent.type = "PluginCenterNotAvailableEvent";
